+++ forth_src/turnkey.fs
@@ -5,8 +5,8 @@ $9fff value oldtop
 start @ value oldstart
 
 : top! ( addr -- )
-latest @ swap top latest @ - 
-2dup - latest ! over to top
+latest swap top latest - 
+2dup - to latest over to top
 swap over - swap 1+ move ;
 
 : restore-forth
@@ -19,9 +19,9 @@ execute ;
 start @ to oldstart
 top to oldtop 
 ['] restore-forth start ! 
-here $20 + top latest @ - + top!
+here $20 + top latest - + top!
 $801 top 1+ $d word count saveb ;
 
 : save-prg ( strptr strlen -- )
-here 0 , top latest ! top!
+here 0 , top to latest top!
 save-pack ;